requirements.txt
Django==2.1.3
django-bootstrap4==0.0.7
django-configurations==2.1
django-debug-toolbar==1.10.1
django-extensions==2.1.4
django-rq==1.2.0
